refactoring SongMeta: all time units in milliseconds and as double
- UltraStarSongMeta handles conversion to txt file specific units

+++ UltraStar Play/Assets/Scenes/SongEditor/Actions/ChangeBpmAction.cs
@@ -46,7 +46,7 @@ public class ChangeBpmAction : INeedInjection
         {
             DivideBpm(voice, factor);
         }
-        float newBpm = songMeta.BeatsPerMinute / factor;
+        double newBpm = songMeta.BeatsPerMinute / factor;
         songMeta.BeatsPerMinute = newBpm;
     }
 
@@ -86,7 +86,7 @@ public class ChangeBpmAction : INeedInjection
         {
             MultiplyBpm(voice, factor);
         }
-        float newBpm = songMeta.BeatsPerMinute * factor;
+        double newBpm = songMeta.BeatsPerMinute * factor;
         songMeta.BeatsPerMinute = newBpm;
     }
